Automatically set GOMAXPROCS based on CPU quota
Introduces `environment.SetMatProcsToCPUQuota` function that uses uber's automaxprocs to automatically
determine the values to set `GOMAXPROCS` to based on the linux container CPU quota. It is enabled
by default when using the `app` package and is disabled via the `AUTO_MAX_PROCS_DISABLED` environment
variable.
